ffmepg-aml: CB2 seek slowly for special ts file [1/1]
PD#SWPL-115243

Problem:
For some special ts streams, the demux did not parse the flag of the
keyframe, resulting in reading to the end after seek.

Solution:
When ffmpeg mpegts seek, it does not parse the key frame flag.
AmFFmpegAdapter is already compatible with the situation that the first
packet readed is not a key frame.

Verify:
AndroidT T5M

Change-Id: Ib79371eb625488b042a4d02909b198d222483672
Signed-off-by: Zhao Yi <zhao.yi@amlogic.com>